edge/inference-service/pyproject.toml
[project]
name = "edge-inference-service"
version = "0.1.0"
description = "Edge gateway inference microservice for MAP prediction"
readme = "README.md"
requires-python = ">=3.11"
authors = [
  { name = "Clinical Automation Team" }
]
dependencies = [
  "fastapi==0.110.0",
  "uvicorn==0.29.0",
  "pydantic==2.6.3",
  "pydantic-settings==2.2.1",
  "grpcio==1.63.0",
  "protobuf==5.29.6",
  "onnxruntime==1.17.1",
  "numpy==1.26.4",
  "httpx==0.27.0",
  "python-dotenv==1.0.1"
]

[project.optional-dependencies]
test = [
  "pytest==8.1.1",
  "pytest-asyncio==0.23.5",
  "respx==0.20.2"
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
